requirements.txt
discord.py>=2.3.0
aiohttp
python-dateutil
pytz
Flask